Minor unrelated cleanup
